Use common to store utils function
