+++ xinference/client/handlers.py
@@ -1,3 +1,21 @@
+from .restful.async_restful_client import (  # noqa: F401
+    AsyncRESTfulAudioModelHandle as AsyncAudioModelHandle,
+)
+from .restful.async_restful_client import (  # noqa: F401
+    AsyncRESTfulChatModelHandle as AsyncChatModelHandle,
+)
+from .restful.async_restful_client import (  # noqa: F401
+    AsyncRESTfulEmbeddingModelHandle as AsyncEmbeddingModelHandle,
+)
+from .restful.async_restful_client import (  # noqa: F401
+    AsyncRESTfulGenerateModelHandle as AsyncGenerateModelHandle,
+)
+from .restful.async_restful_client import (  # noqa: F401
+    AsyncRESTfulImageModelHandle as AsyncImageModelHandle,
+)
+from .restful.async_restful_client import (  # noqa: F401
+    AsyncRESTfulVideoModelHandle as AsyncVideoModelHandle,
+)
 from .restful.restful_client import (  # noqa: F401
     RESTfulAudioModelHandle as AudioModelHandle,
 )
