+++ xinference/client/restful/async_restful_client.py
@@ -17,8 +17,7 @@ from typing import TYPE_CHECKING, Any, AsyncIterator, Dict, List, Optional, Unio
 
 import aiohttp
 
-from ..common import async_streaming_response_iterator
-from ..utils import convert_float_to_int_or_str
+from ..common import async_streaming_response_iterator, convert_float_to_int_or_str
 
 if TYPE_CHECKING:
     from ...types import (
